btrfs-progs: use libbtrfsutil for subvolume list
This is the most involved conversion because it replaces the libbtrfs
implementation entirely. I took care to preserve the existing behavior
in all cases, warts and all.

This also moves the list printing code from libbtrfs to
cmds-subvolume.c. This is an ABI break, but it avoids adding a
libbtrfsutil dependency to libbtrfs. A search online didn't turn up any
projects which use this interface.

Signed-off-by: Omar Sandoval <osandov@fb.com>