Btrfs-progs: send, bump stream version
This increases the send stream version from version 1 to version 2, adding
new commands:

1) fallocate - used to pre-allocate space for files and to punch holes in files

2) FS_IOC_SETFLAGS (a.k.a. chattr)

3) send compressed write

This is preparation work for subsequent changes that implement the new
features.

This doesn't break compatibility with older kernels or clients. In order
to get a version 2 send stream, a new flag must be passed to the send
ioctl.

Signed-off-by: Filipe David Borba Manana <fdmanana@gmail.com>
[Omar: drop total data size and utimes2 for now]
Signed-off-by: Omar Sandoval <osandov@fb.com>