Btrfs-progs: send, implement fallocate command callback
The fallocate send stream command, added in stream version 2, is used to
pre-allocate space for files and punch file holes. This change
implements the callback for that new command, using the fallocate
function from the standard C library to carry out the specified action
(allocate file space or punch a file hole).

Signed-off-by: Filipe David Borba Manana <fdmanana@gmail.com>
[Omar: add fallocate command to send dump]
Signed-off-by: Omar Sandoval <osandov@fb.com>